Created Teleport
Created a working teleport, along with changing the UI to represent when you can teleport. I changed it a bit from the original idea: There is now a teleporter object that when you touch, lets you teleport anywhere. This allows for more creativity and I think is more fun than the original idea.